Домашнее задание 6: JPQL, Spring ORM, DAO на основе Spring ORM + JPA

+++ spring-06/src/main/java/ru/otus/spring/services/CommentServiceImpl.java
@@ -35,7 +35,9 @@ public class CommentServiceImpl implements CommentService {
     @Transactional
     public Comment removeComment(long id) {
         Comment comment  = commentRepository.findById(id);
-        commentRepository.remove(comment);
+        if(comment != null){
+            commentRepository.remove(comment);
+        }
         return comment;
     }
